Add update flag to tuist fetch
