--- Sources/TuistKit/Commands/DependenciesCommand.swift
@@ -1,15 +1,0 @@
-import ArgumentParser
-import Foundation
-import TSCBasic
-
-struct DependenciesCommand: ParsableCommand {
-    static var configuration: CommandConfiguration {
-        CommandConfiguration(
-            commandName: "dependencies",
-            abstract: "A set of commands for dependencies' management.",
-            subcommands: [
-                DependenciesUpdateCommand.self,
-            ]
-        )
-    }
-}
